* Increased Clang warning level (-Weverything) and fixed the resulting warnings. Most were innocuous, but it did turn up some legitimate problems/concerns which were also fixed (such as HadesMemError not having an 'anchored' vtable, and also some integer conversion 'problems'). * Worked around a Clang codegen bug (caused by catch-all in EH) in memhack.

+++ include/memory/hadesmem/process_entry.hpp
@@ -51,7 +51,7 @@ public:
     return parent_;
   }
 
-  DWORD GetPriority() const BOOST_NOEXCEPT
+  LONG GetPriority() const BOOST_NOEXCEPT
   {
     return priority_;
   }
@@ -65,7 +65,7 @@ private:
   DWORD id_;
   DWORD threads_;
   DWORD parent_;
-  DWORD priority_;
+  LONG priority_;
   std::wstring name_;
 };
